Refactor: doble confirmacion en archivo js y eliminar archivos basura, Fix: Seeder.
